+++ src/components/Navigationbar.js
@@ -10,9 +10,9 @@ import {
   Button
   } from 'reactstrap';
 
-  import * as Scroll from 'react-scroll';
-  import { Link, Element , Events, animateScroll as scroll, scrollSpy, scroller } from 'react-scroll'
-   
+import * as Scroll from 'react-scroll';
+import { Link, Element , Events, animateScroll as scroll, scrollSpy, scroller } from 'react-scroll'
+import './styles/navigation-default.css';
 
 export default class Navigationbar extends React.Component {
   constructor(props) {
@@ -35,12 +35,18 @@ export default class Navigationbar extends React.Component {
 
   gotoAbout() {
     var html = document.documentElement.clientHeight;
-    
+
     scroll.scrollTo(html);
   }
-  
+
   gotoProjects() {
     var html = document.documentElement.clientHeight * 2;
+
+    scroll.scrollTo(html);
+  }
+
+  gotoContact() {
+    var html = document.documentElement.clientHeight * 3;
 
     scroll.scrollTo(html);
   }
@@ -48,16 +54,19 @@ export default class Navigationbar extends React.Component {
   render() {
     return (
       <div>
-        <Navbar color="none" dark expand="md" fixed={`top`} className="navbar-border">
-          <NavbarBrand className="brand" onClick={this.gotoTop}><b>coding</b>mroberts</NavbarBrand>
+        <Navbar color="none" dark expand="md" fixed={`top`} id="navbar-border">
+          <NavbarBrand id="brand" onClick={this.gotoTop}><b>coding</b>mroberts</NavbarBrand>
           <NavbarToggler onClick={this.toggle} />
           <Collapse isOpen={this.state.isOpen} navbar>
             <Nav className="ml-auto" navbar>
-              <NavItem>
-                <NavLink className="nav-button" onClick={this.gotoAbout}>About</NavLink>
+              <NavItem id="nav-item">
+                <Button id="nav-button" color="link"onClick={this.gotoAbout}>About</Button>
               </NavItem>
               <NavItem>
-                <NavLink className="nav-button" onClick={this.gotoProjects}>Projects</NavLink>
+                <Button id="nav-button" color="link"onClick={this.gotoProjects}>Projects</Button>
+              </NavItem>
+              <NavItem>
+                <Button id="nav-button" color="link"onClick={this.gotoContact}>Contact</Button>
               </NavItem>
             </Nav>
           </Collapse>
