+++ src/components/styles/navigation-default.css
@@ -1,0 +1,22 @@
+#navbar-border {
+  border: 0;
+  background-image: linear-gradient(to right, rgba(255, 255, 255, 0.13), rgba(255, 255, 255, 0.0), rgba(255, 255, 255, 0.13));
+  box-shadow: 0 0.161rem 0.08rem rgba(15, 10, 10, 0.162);
+}
+#nav-button:hover {
+  padding-top: 10px;
+}
+
+#nav-button {
+  text-decoration: none;
+  color: rgb(17, 150, 150);
+}
+
+#brand {
+  cursor: pointer;
+}
+
+#brand:hover {
+  cursor: pointer;
+  padding-top: 10px;
+}
