--- src/pages/styles/aboutme-default.css
@@ -1,15 +1,0 @@
-#aboutme-view {
-  height: 100vh;
-}
-#img {
-  width: 100%;
-  padding: 10px;
-}
-#section-title {
-  padding-left: 4vw;
-}
-#section-text {
-  padding-right: 3vw;
-  background-color: red;
-  padding-top: 10vh;
-}
