+++ src/pages/styles/landing-default.css
@@ -5,13 +5,24 @@
   justify-content: center;
 }
 #align-center {
+  display:flex;
+  flex-direction: column;
   text-align: center;
-  border: 3px green solid;
+  align-items: center;
+  justify-content: center;
   max-width: 75%;
 }
 #social-but-links {
   padding-top: 3vmax;
   flex-wrap: wrap;
+}
+#connect-button {
+  color: rgb(17, 150, 150);
+  border-color: rgb(17, 150, 150);
+}
+#connect-button:hover {
+  color: white;
+  background-color: rgb(17, 150, 150);
 }
 #welcome-info-body {
   margin-top: 7vmax;
@@ -22,7 +33,7 @@
 #title {
   font-size: 6.93vmax;
   font-family: Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif;
-  color: rgb(102, 252, 241);
+  color: rgb(17, 150, 150);
 }
 #subtitle {
   font-size: 3.69vmax;
